import_paths managed by the app
