+++ tests/unit/test_worker.py
@@ -1,5 +1,3 @@
-import sys
-
 import pendulum
 import pytest
 
@@ -225,20 +223,13 @@ def test_run_job_not_found(app, job_store):
         test_worker.run_job(job)
 
 
-def test_import_all():
-    module = "tests.unit.unused_module"
-    assert module not in sys.modules
-
-    worker.import_all([module])
-
-    assert module in sys.modules
-
-
 def test_worker_call_import_all(app, mocker):
 
-    import_all = mocker.patch("procrastinate.worker.import_all")
+    import_all = mocker.patch("procrastinate.utils.import_all")
 
-    worker.Worker(app=app, queues=["yay"], import_paths=["hohoho"])
+    app.import_paths = ["hohoho"]
+
+    worker.Worker(app=app, queues=["yay"])
 
     import_all.assert_called_with(import_paths=["hohoho"])
